Usability Update: Interface Features
For both the MSFVenom payload generators and custom nmapping, there now exists dictionaries in the code that contain every type of encoder and required syntax, and scan parameter that is needed.

All you have to do is press a button (a number) to select your encoder as well as your desired meterpreter payload as it prints out on the list. No need to struggle with leaving out the backslashes on 'set PAYLOAD windows/meterpreter/reverse_https' anymore.

Same thing for customized NMap Scans. You don't have to "hit the ManPage" anymore. Pick a number from the list to choose from a FIN scan all the way to a FTP bounce or Zombie Scan. 

Check the ReadMe for new upcoming features that will be added. (We are considering a few handy tools, tips and tricks offered by Violent Python and Black Hat Python).